Add a max width to the calculations
