+++ app/src/main/java/com/github/damontecres/stashapp/presenters/PopupOnLongClickListener.kt
@@ -1,15 +1,12 @@
 package com.github.damontecres.stashapp.presenters
 
-import android.content.Context
 import android.view.View
-import android.view.View.MeasureSpec
 import android.view.View.OnLongClickListener
-import android.widget.Adapter
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
-import android.widget.FrameLayout
 import androidx.appcompat.widget.ListPopupWindow
 import com.github.damontecres.stashapp.R
+import com.github.damontecres.stashapp.util.getMaxMeasuredWidth
 
 /**
  * An OnLongClickListener which shows a popup of predefined options
@@ -37,7 +34,7 @@ class PopupOnLongClickListener(
         listPopUp.anchorView = view
         listPopUp.width =
             if (popUpWidth == ListPopupWindow.WRAP_CONTENT) {
-                getMaxWidth(view.context, adapter)
+                getMaxMeasuredWidth(view.context, adapter)
             } else {
                 popUpWidth
             }
@@ -53,30 +50,5 @@ class PopupOnLongClickListener(
         listPopUp.show()
 
         return true
-    }
-
-    private fun getMaxWidth(
-        context: Context,
-        adapter: ArrayAdapter<String>,
-    ): Int {
-        if (adapter.viewTypeCount != 1) {
-            throw IllegalStateException("Adapter creates more than 1 type of view")
-        }
-
-        val tempParent = FrameLayout(context)
-        var maxWidth = 0
-        var itemView: View? = null
-        val measureSpec = MeasureSpec.makeMeasureSpec(0, MeasureSpec.UNSPECIFIED)
-
-        for (i in 0 until adapter.count) {
-            if (adapter.getItemViewType(i) != Adapter.IGNORE_ITEM_VIEW_TYPE) {
-                itemView = adapter.getView(i, itemView, tempParent)
-                itemView.measure(measureSpec, measureSpec)
-                if (itemView.measuredWidth > maxWidth) {
-                    maxWidth = itemView.measuredWidth
-                }
-            }
-        }
-        return maxWidth
     }
 }
